fix: Rename resources_reservation module to resourcesReservation and reserve resources after calculating missing resources
